third_party/blink/web_tests/fast/harness/expected-error-in-async-test-expected.txt

CONSOLE ERROR: Uncaught Error: An expected error
Test that the window.onerror handler in resources/js-test.js stops asynchronous tests

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Will throw an error soon
Throwing an expected error
PASS Got expected error: 'expected error'
PASS successfullyParsed is true

TEST COMPLETE
